+++ services/src/main/java/by/epam/tc/conference/services/impl/ProposalServiceImpl.java
@@ -2,36 +2,36 @@ package by.epam.tc.conference.services.impl;
 
 import by.epam.tc.conference.dao.DaoException;
 import by.epam.tc.conference.dao.ProposalDao;
-import by.epam.tc.conference.dao.ProposalDetailsDao;
-import by.epam.tc.conference.dto.ProposalDetails;
+import by.epam.tc.conference.dao.UserDao;
 import by.epam.tc.conference.entity.Proposal;
 import by.epam.tc.conference.entity.ProposalStatus;
+import by.epam.tc.conference.entity.User;
 import by.epam.tc.conference.services.ProposalService;
-import by.epam.tc.conference.services.exception.EntityNotFoundException;
-import by.epam.tc.conference.services.exception.InvalidEntityException;
+import by.epam.tc.conference.services.exception.NoAuthorityException;
+import by.epam.tc.conference.services.exception.NotFoundException;
+import by.epam.tc.conference.services.exception.InvalidDataException;
 import by.epam.tc.conference.services.exception.ServiceException;
 import by.epam.tc.conference.services.validator.Validator;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
-import java.util.List;
 import java.util.Optional;
 
 public class ProposalServiceImpl implements ProposalService {
 
     private static final Logger logger = LogManager.getLogger(ProposalServiceImpl.class);
     private final ProposalDao proposalDao;
-    private final ProposalDetailsDao proposalDetailsDao;
+    private final UserDao userDao;
     private final Validator<Proposal> validator;
 
-    public ProposalServiceImpl(ProposalDao proposalDao, ProposalDetailsDao proposalDetailsDao, Validator<Proposal> validator) {
+    public ProposalServiceImpl(ProposalDao proposalDao, UserDao userDao, Validator<Proposal> validator) {
         this.proposalDao = proposalDao;
-        this.proposalDetailsDao = proposalDetailsDao;
+        this.userDao = userDao;
         this.validator = validator;
     }
 
     @Override
-    public void createProposal(Proposal proposal) throws ServiceException, InvalidEntityException {
+    public void createProposal(Proposal proposal) throws ServiceException {
         try {
             validateProposal(proposal);
             proposalDao.save(proposal);
@@ -42,8 +42,11 @@ public class ProposalServiceImpl implements ProposalService {
     }
 
     @Override
-    public void deleteProposal(Long id) throws ServiceException {
+    public void deleteProposal(long id, long userId) throws ServiceException {
         try {
+            if (!hasAuthority(id, userId)) {
+                throw new NoAuthorityException("Cannot delete proposal id=" + id + " No authority");
+            }
             proposalDao.deleteById(id);
             logger.info("Deleted proposal id={}", id);
         } catch (DaoException e) {
@@ -52,33 +55,11 @@ public class ProposalServiceImpl implements ProposalService {
     }
 
     @Override
-    public List<ProposalDetails> findProposalsDetailsBySectionId(Long id) throws ServiceException {
+    public Proposal getProposal(long proposalId) throws ServiceException {
         try {
-            List<ProposalDetails> proposals = proposalDetailsDao.findProposalsBySectionId(id);
-            logger.debug("Found {} details for proposals in section with id={}", proposals.size(), id);
-            return proposals;
-        } catch (DaoException e) {
-            throw new ServiceException(e.getMessage(), e);
-        }
-    }
-
-    @Override
-    public List<ProposalDetails> findProposalsDetailsByParticipantId(Long id) throws ServiceException {
-        try {
-            List<ProposalDetails> proposals = proposalDetailsDao.findProposalsByUserId(id);
-            logger.debug("Found {} details for proposals with participantId={}", proposals.size(), id);
-            return proposals;
-        } catch (DaoException e) {
-            throw new ServiceException(e.getMessage(), e);
-        }
-    }
-
-    @Override
-    public Proposal getProposal(Long id) throws ServiceException, EntityNotFoundException {
-        try {
-            Optional<Proposal> optionalProposal = proposalDao.findById(id);
+            Optional<Proposal> optionalProposal = proposalDao.findById(proposalId);
             Proposal proposal = optionalProposal.orElseThrow(() ->
-                    new EntityNotFoundException("There is no conference with id=" + id)
+                    new NotFoundException("There is no conference with id=" + proposalId)
             );
             logger.info("conference with id={} returned");
             return proposal;
@@ -88,8 +69,11 @@ public class ProposalServiceImpl implements ProposalService {
     }
 
     @Override
-    public void updateStatus(Long id, ProposalStatus status) throws ServiceException {
+    public void updateStatus(long id, ProposalStatus status, long userId) throws ServiceException {
         try {
+            if (!canChangeStatus(id, userId)) {
+                throw new NoAuthorityException("Cannot change status for proposal id=" + id + " No authority");
+            }
             proposalDao.updateStatus(id, status);
             logger.info("Changed status for proposal id={} to {}", id, status);
         } catch (DaoException e) {
@@ -98,9 +82,13 @@ public class ProposalServiceImpl implements ProposalService {
     }
 
     @Override
-    public void updateProposal(Proposal proposal) throws ServiceException, InvalidEntityException {
+    public void updateProposal(Proposal proposal, long userId) throws ServiceException {
         try {
             validateProposal(proposal);
+            Long proposalId = proposal.getId();
+            if (!hasAuthority(proposalId, userId)) {
+                throw new NoAuthorityException("Cannot update proposal id=" + proposalId + " No authority");
+            }
             proposalDao.update(proposal);
             logger.info("Updated proposal id={}", proposal.getId());
         } catch (DaoException e) {
@@ -108,11 +96,36 @@ public class ProposalServiceImpl implements ProposalService {
         }
     }
 
-    private void validateProposal(Proposal proposal) throws InvalidEntityException {
+    /**
+     * When proposal is invalid throws {@link InvalidDataException}
+     * Proposal id == null is valid state
+     * @param proposal to validate
+     * @throws InvalidDataException when proposal invalid
+     */
+    private void validateProposal(Proposal proposal) throws InvalidDataException {
         boolean isValid = validator.validate(proposal);
         if (!isValid) {
-            throw new InvalidEntityException("Invalid proposal id=" + proposal.getId());
+            throw new InvalidDataException("Invalid proposal id=" + proposal.getId());
         }
     }
 
+    private boolean canChangeStatus(long proposalId, long userId) throws DaoException {
+        Optional<User> optionalAdministrator = userDao.findAdministratorByProposalId(proposalId);
+        if (!optionalAdministrator.isPresent()) {
+            return false;
+        }
+        User administrator = optionalAdministrator.get();
+        Long administratorId = administrator.getId();
+        return administratorId == userId;
+    }
+
+    private boolean hasAuthority(long proposalId, long userId) throws DaoException {
+        Optional<Proposal> optionalProposal = proposalDao.findById(proposalId);
+        if (!optionalProposal.isPresent()) {
+            return false;
+        }
+        Proposal proposal = optionalProposal.get();
+        Long participantId = proposal.getParticipantId();
+        return userId == participantId;
+    }
 }
